LMS/lms_backend/requirements.txt
acres==0.5.0
aiofiles==24.1.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
boto3==1.40.17
botocore==1.40.17
build==1.3.0
cachetools==6.2.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
chromadb==1.3.0
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
docutils==0.22
durationpy==0.10
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
etelemetry==0.3.1
executing==2.2.0
faiss-cpu==1.12.0
fastapi==0.116.1
fastapi-cli==0.0.8
filelock==3.19.1
fitz==0.0.1.dev2
flatbuffers==25.9.23
fonttools==4.60.0
frontend==0.0.3
fsspec==2025.9.0
google-auth==2.42.0
googleapis-common-protos==1.71.0
greenlet==3.2.3
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
id==1.5.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jpype1==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.6.0
kiwisolver==1.4.9
konlpy==0.6.0
looseversion==1.3.0
lxml==6.0.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplot==0.1.9
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.2.0
more-itertools==10.8.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
nh3==0.3.0
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.23.2
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.2
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.3.2
parso==0.8.4
passlib==1.7.4
pathlib==1.0.1
pillow==11.3.0
platformdirs==4.3.7
posthog==5.4.0
prompt_toolkit==3.0.50
protobuf==6.33.0
prov==2.1.1
psutil==7.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
pure_eval==0.2.3
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.1
PyJWT==2.10.1
pyloco==0.0.139
PyMuPDF==1.26.4
PyMuPDFb==1.24.9
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.14.0
rdflib==7.1.4
readme_renderer==44.0
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.1.0
rich-toolkit==0.15.0
rpds-py==0.28.0
rsa==4.9.1
s3transfer==0.13.1
safetensors==0.6.2
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
sentence-transformers==5.1.2
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.1
SimpleWebSocketServer==0.1.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.47.3
sympy==1.14.0
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.57.1
trio==0.30.0
trio-websocket==0.12.2
twine==6.2.0
typer==0.16.1
typing==3.7.4.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.11.0
Unidecode==1.4.0
urllib3
ushlex==0.99.1
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
xlrd==2.0.2
zipp==3.23.0
